+++ packages/components/src/components/button/button.tsx
@@ -41,6 +41,8 @@ export class Button {
   @Prop() target?: string = '_self';
   /** (optional) Injected CSS styles */
   @Prop() styles?: string;
+  /** (optional) If `true`, a download is triggrered */
+  @Prop() download?: boolean = false;
 
   /**
    * Prevent clicks from being emitted from the host
@@ -85,7 +87,10 @@ export class Button {
    * If so, it's probably an icon, so we set `iconPosition` to `after`.
    */
   setIconPositionProp() {
-    const nodes = Array.from(this.hostElement.childNodes);
+    const nodes = Array.from(this.hostElement.childNodes).filter((node) => {
+      // ignore empty text nodes, which are probably due to formatting
+      return !(node.nodeType === 3 && node.nodeValue.trim() === '');
+    });
     if (nodes.length < 2) {
       return;
     }
@@ -112,6 +117,7 @@ export class Button {
           <a
             class={this.getCssClassMap()}
             href={this.href}
+            download={this.download}
             target={this.target}
             rel={this.target === '_blank' ? 'noopener noreferrer' : undefined}
             aria-label={this.ariaLabel}