+++ packages/components/src/components/text-field/text-field.tsx
@@ -169,7 +169,7 @@ export class TextField {
             disabled={this.disabled}
             readonly={this.readonly}
             {...ariaInvalidAttr}
-            {...ariaDescribedByAttr}
+            {...(this.helperText ? ariaDescribedByAttr : {})}
           />
 
           {(!!this.helperText || !!this.counter) && (
@@ -208,6 +208,7 @@ export class TextField {
       this.transparent && 'text-field--transparent',
       this.status && `text-field--status-${this.status}`,
       this.size && `text-field--size-${this.size}`,
+      this.readonly && `text-field--readonly`,
       animated && 'animated'
     );
   }