+++ packages/storybook-vue/.storybook/preview.js
@@ -1,9 +1,9 @@
-import Vue from "vue";
+import Vue from 'vue';
 
 import {
   defineCustomElements,
-  applyPolyfills
-} from "@telekom/scale-components/loader";
+  applyPolyfills,
+} from '@telekom/scale-components/loader';
 
 Vue.config.productionTip = false;
 Vue.config.ignoredElements = [/scale-\w*/];
@@ -11,3 +11,24 @@ Vue.config.ignoredElements = [/scale-\w*/];
 applyPolyfills().then(() => {
   defineCustomElements(window);
 });
+
+export const parameters = {
+  a11y: {
+    config: {
+      rules: [
+        {
+          id: 'list',
+          enabled: false,
+        },
+        {
+          id: 'listitem',
+          enabled: false,
+        },
+        {
+          id: 'color-contrast',
+          enabled: false,
+        },
+      ],
+    },
+  },
+};