+++ packages/storybook-vue/stories/3_components/accordion/ScaleAccordion.vue
@@ -1,5 +1,5 @@
 <template>
-  <scale-accordion :styles="styles" :dependent="dependent" :expanded="expanded">
+  <scale-accordion :styles="styles" :dependent="dependent" :expanded="expanded" :heading-level="headingLevel">
     <slot></slot>
   </scale-accordion>
 </template>
@@ -10,6 +10,7 @@ export default {
     styles: { type: String },
     dependent: { type: Boolean, default: false },
     expanded: { type: Boolean, default: false },
+    headingLevel: { type: Number },
   },
 };
 </script>