+++ packages/visual-tests/src/checkbox.visual.spec.js
@@ -4,8 +4,9 @@ describe('Checkbox', () => {
     ['standard-disabled'],
     ['selected'],
     ['selected-disabled'],
+    ['helper-text'],
     ['error'],
-    ['custom-label-slot'],
+    ['custom-label'],
   ])('%p', async (variant) => {
     await global.page.goto(
       `http://host.docker.internal:3123/iframe.html?id=components-checkbox--${variant}&viewMode=story`