+++ packages/components-vue/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-beta.16](https://github.com/telekom/scale/compare/v3.0.0-beta.15...v3.0.0-beta.16) (2021-07-20)
+
+**Note:** Version bump only for package @telekom/scale-components-vue
+
+
+
+
+
+# [3.0.0-beta.15](https://github.com/telekom/scale/compare/v3.0.0-beta.14...v3.0.0-beta.15) (2021-06-21)
+
+**Note:** Version bump only for package @telekom/scale-components-vue
+
+
+
+
+
+# [3.0.0-beta.14](https://github.com/telekom/scale/compare/v3.0.0-beta.13...v3.0.0-beta.14) (2021-06-15)
+
+**Note:** Version bump only for package @telekom/scale-components-vue
+
+
+
+
+
+# [3.0.0-beta.13](https://github.com/telekom/scale/compare/v3.0.0-beta.12...v3.0.0-beta.13) (2021-06-07)
+
+**Note:** Version bump only for package @telekom/scale-components-vue
+
+
+
+
+
+# [3.0.0-beta.12](https://github.com/telekom/scale/compare/v3.0.0-beta.11...v3.0.0-beta.12) (2021-05-27)
+
+**Note:** Version bump only for package @telekom/scale-components-vue
+
+
+
+
+
+# [3.0.0-beta.11](https://github.com/telekom/scale/compare/v3.0.0-beta.10...v3.0.0-beta.11) (2021-05-26)
+
+**Note:** Version bump only for package @telekom/scale-components-vue
+
+
+
+
+
+# [3.0.0-beta.10](https://github.com/telekom/scale/compare/v3.0.0-beta.9...v3.0.0-beta.10) (2021-05-25)
+
+**Note:** Version bump only for package @telekom/scale-components-vue
+
+
+
+
+
+# [3.0.0-beta.9](https://github.com/telekom/scale/compare/v3.0.0-beta.8...v3.0.0-beta.9) (2021-05-17)
+
+**Note:** Version bump only for package @telekom/scale-components-vue
+
+
+
+
+
+# [3.0.0-beta.8](https://github.com/telekom/scale/compare/v3.0.0-beta.7...v3.0.0-beta.8) (2021-05-12)
+
+**Note:** Version bump only for package @telekom/scale-components-vue
+
+
+
+
+
+# [3.0.0-beta.7](https://github.com/telekom/scale/compare/v3.0.0-beta.6...v3.0.0-beta.7) (2021-05-06)
+
+**Note:** Version bump only for package @telekom/scale-components-vue
+
+
+
+
+
+# [3.0.0-beta.6](https://github.com/telekom/scale/compare/v3.0.0-beta.5...v3.0.0-beta.6) (2021-05-03)
+
+**Note:** Version bump only for package @telekom/scale-components-vue
+
+
+
+
+
 # [3.0.0-beta.5](https://github.com/telekom/scale/compare/v3.0.0-beta.4...v3.0.0-beta.5) (2021-04-28)
 
 **Note:** Version bump only for package @telekom/scale-components-vue