+++ packages/html-to-sketch/CHANGELOG.md
@@ -3,6 +3,100 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-beta.16](https://github.com/telekom/scale/compare/v3.0.0-beta.15...v3.0.0-beta.16) (2021-07-20)
+
+
+### Bug Fixes
+
+* format ([#453](https://github.com/telekom/scale/issues/453)) ([de170f7](https://github.com/telekom/scale/commit/de170f7e159e7bc05a804b0e391dc7d3b7af6c3f))
+
+
+
+
+
+# [3.0.0-beta.15](https://github.com/telekom/scale/compare/v3.0.0-beta.14...v3.0.0-beta.15) (2021-06-21)
+
+**Note:** Version bump only for package @telekom/scale-html-to-sketch
+
+
+
+
+
+# [3.0.0-beta.14](https://github.com/telekom/scale/compare/v3.0.0-beta.13...v3.0.0-beta.14) (2021-06-15)
+
+**Note:** Version bump only for package @telekom/scale-html-to-sketch
+
+
+
+
+
+# [3.0.0-beta.13](https://github.com/telekom/scale/compare/v3.0.0-beta.12...v3.0.0-beta.13) (2021-06-07)
+
+**Note:** Version bump only for package @telekom/scale-html-to-sketch
+
+
+
+
+
+# [3.0.0-beta.12](https://github.com/telekom/scale/compare/v3.0.0-beta.11...v3.0.0-beta.12) (2021-05-27)
+
+**Note:** Version bump only for package @telekom/scale-html-to-sketch
+
+
+
+
+
+# [3.0.0-beta.11](https://github.com/telekom/scale/compare/v3.0.0-beta.10...v3.0.0-beta.11) (2021-05-26)
+
+**Note:** Version bump only for package @telekom/scale-html-to-sketch
+
+
+
+
+
+# [3.0.0-beta.10](https://github.com/telekom/scale/compare/v3.0.0-beta.9...v3.0.0-beta.10) (2021-05-25)
+
+
+### Features
+
+* **sketch-generator:** add icon merging to generator ([#324](https://github.com/telekom/scale/issues/324)) ([3b7cbd6](https://github.com/telekom/scale/commit/3b7cbd6fbb92393a408126601c909ee6b4a1e89f))
+
+
+
+
+
+# [3.0.0-beta.9](https://github.com/telekom/scale/compare/v3.0.0-beta.8...v3.0.0-beta.9) (2021-05-17)
+
+**Note:** Version bump only for package @telekom/scale-html-to-sketch
+
+
+
+
+
+# [3.0.0-beta.8](https://github.com/telekom/scale/compare/v3.0.0-beta.7...v3.0.0-beta.8) (2021-05-12)
+
+**Note:** Version bump only for package @telekom/scale-html-to-sketch
+
+
+
+
+
+# [3.0.0-beta.7](https://github.com/telekom/scale/compare/v3.0.0-beta.6...v3.0.0-beta.7) (2021-05-06)
+
+**Note:** Version bump only for package @telekom/scale-html-to-sketch
+
+
+
+
+
+# [3.0.0-beta.6](https://github.com/telekom/scale/compare/v3.0.0-beta.5...v3.0.0-beta.6) (2021-05-03)
+
+**Note:** Version bump only for package @telekom/scale-html-to-sketch
+
+
+
+
+
 # [3.0.0-beta.5](https://github.com/telekom/scale/compare/v3.0.0-beta.4...v3.0.0-beta.5) (2021-04-28)
 
 **Note:** Version bump only for package @telekom/scale-html-to-sketch