+++ packages/storybook-vue/CHANGELOG.md
@@ -3,6 +3,119 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-beta.16](https://github.com/telekom/scale/compare/v3.0.0-beta.15...v3.0.0-beta.16) (2021-07-20)
+
+
+### Features
+
+* add documentation for new Actions data-grid cell type ([#464](https://github.com/telekom/scale/issues/464)) ([5f9f6ef](https://github.com/telekom/scale/commit/5f9f6ef04c20a8892b55850017e65bfeea6937ef))
+* badge ([#394](https://github.com/telekom/scale/issues/394)) ([686ac2d](https://github.com/telekom/scale/commit/686ac2d6f366e2676ad0d96abce2714a02c334e1))
+* implement scaleBeforeClose event in Modal ([#430](https://github.com/telekom/scale/issues/430)) ([bdc6786](https://github.com/telekom/scale/commit/bdc678655e4d3921ab546448cca70f38ae4057ee))
+
+
+
+
+
+# [3.0.0-beta.15](https://github.com/telekom/scale/compare/v3.0.0-beta.14...v3.0.0-beta.15) (2021-06-21)
+
+
+### Bug Fixes
+
+* **a11y:** axe audit ([#421](https://github.com/telekom/scale/issues/421)) ([39a0021](https://github.com/telekom/scale/commit/39a002141aa909eb1100bc5a8f727e0a1ba462a0))
+
+
+### Features
+
+* add fill prop to icon with higher priority than color ([#412](https://github.com/telekom/scale/issues/412)) ([4f4bac4](https://github.com/telekom/scale/commit/4f4bac41915b0c64d6f43ca91eef6254c08bb29f))
+
+
+
+
+
+# [3.0.0-beta.14](https://github.com/telekom/scale/compare/v3.0.0-beta.13...v3.0.0-beta.14) (2021-06-15)
+
+
+### Features
+
+* **button:** add download attribute ([#380](https://github.com/telekom/scale/issues/380)) ([49b14c0](https://github.com/telekom/scale/commit/49b14c0e1c211f7fa882943d2c82ab93699cc7f0))
+
+
+
+
+
+# [3.0.0-beta.13](https://github.com/telekom/scale/compare/v3.0.0-beta.12...v3.0.0-beta.13) (2021-06-07)
+
+**Note:** Version bump only for package @telekom/scale-storybook-vue
+
+
+
+
+
+# [3.0.0-beta.12](https://github.com/telekom/scale/compare/v3.0.0-beta.11...v3.0.0-beta.12) (2021-05-27)
+
+**Note:** Version bump only for package @telekom/scale-storybook-vue
+
+
+
+
+
+# [3.0.0-beta.11](https://github.com/telekom/scale/compare/v3.0.0-beta.10...v3.0.0-beta.11) (2021-05-26)
+
+**Note:** Version bump only for package @telekom/scale-storybook-vue
+
+
+
+
+
+# [3.0.0-beta.10](https://github.com/telekom/scale/compare/v3.0.0-beta.9...v3.0.0-beta.10) (2021-05-25)
+
+
+### Bug Fixes
+
+* deprecate direct DOM manipulation in Collapsible, fixes [#316](https://github.com/telekom/scale/issues/316) ([06e8e2c](https://github.com/telekom/scale/commit/06e8e2c34e8449b37cdf4b9b52f2131b30dffad2))
+
+
+
+
+
+# [3.0.0-beta.9](https://github.com/telekom/scale/compare/v3.0.0-beta.8...v3.0.0-beta.9) (2021-05-17)
+
+**Note:** Version bump only for package @telekom/scale-storybook-vue
+
+
+
+
+
+# [3.0.0-beta.8](https://github.com/telekom/scale/compare/v3.0.0-beta.7...v3.0.0-beta.8) (2021-05-12)
+
+**Note:** Version bump only for package @telekom/scale-storybook-vue
+
+
+
+
+
+# [3.0.0-beta.7](https://github.com/telekom/scale/compare/v3.0.0-beta.6...v3.0.0-beta.7) (2021-05-06)
+
+
+### Features
+
+* **text-field:** styled read-only attribute w/ tokens ([#313](https://github.com/telekom/scale/issues/313)) ([1abbdd6](https://github.com/telekom/scale/commit/1abbdd6c2bbda810ab3529026d0f837e649ae792))
+
+
+
+
+
+# [3.0.0-beta.6](https://github.com/telekom/scale/compare/v3.0.0-beta.5...v3.0.0-beta.6) (2021-05-03)
+
+
+### Features
+
+* **footer:** dynamic copyright text ([#318](https://github.com/telekom/scale/issues/318)) ([04d1ad3](https://github.com/telekom/scale/commit/04d1ad3b247159e10a51db88ce56e658b42147ed))
+
+
+
+
+
 # [3.0.0-beta.5](https://github.com/telekom/scale/compare/v3.0.0-beta.4...v3.0.0-beta.5) (2021-04-28)