+++ packages/visual-tests/CHANGELOG.md
@@ -3,6 +3,109 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-beta.16](https://github.com/telekom/scale/compare/v3.0.0-beta.15...v3.0.0-beta.16) (2021-07-20)
+
+**Note:** Version bump only for package @telekom/scale-visual-tests
+
+
+
+
+
+# [3.0.0-beta.15](https://github.com/telekom/scale/compare/v3.0.0-beta.14...v3.0.0-beta.15) (2021-06-21)
+
+
+### Bug Fixes
+
+* **a11y:** axe audit ([#421](https://github.com/telekom/scale/issues/421)) ([39a0021](https://github.com/telekom/scale/commit/39a002141aa909eb1100bc5a8f727e0a1ba462a0))
+
+
+
+
+
+# [3.0.0-beta.14](https://github.com/telekom/scale/compare/v3.0.0-beta.13...v3.0.0-beta.14) (2021-06-15)
+
+**Note:** Version bump only for package @telekom/scale-visual-tests
+
+
+
+
+
+# [3.0.0-beta.13](https://github.com/telekom/scale/compare/v3.0.0-beta.12...v3.0.0-beta.13) (2021-06-07)
+
+**Note:** Version bump only for package @telekom/scale-visual-tests
+
+
+
+
+
+# [3.0.0-beta.12](https://github.com/telekom/scale/compare/v3.0.0-beta.11...v3.0.0-beta.12) (2021-05-27)
+
+**Note:** Version bump only for package @telekom/scale-visual-tests
+
+
+
+
+
+# [3.0.0-beta.11](https://github.com/telekom/scale/compare/v3.0.0-beta.10...v3.0.0-beta.11) (2021-05-26)
+
+
+### Features
+
+* **design-tokens:** add functional colors to match current documentation ([#364](https://github.com/telekom/scale/issues/364)) ([a3a0821](https://github.com/telekom/scale/commit/a3a08216e872732ad81a1f555a9a2b65a06c7f97))
+
+
+
+
+
+# [3.0.0-beta.10](https://github.com/telekom/scale/compare/v3.0.0-beta.9...v3.0.0-beta.10) (2021-05-25)
+
+
+### Features
+
+* **date-picker:** flatten top/bottom left border radius ([#355](https://github.com/telekom/scale/issues/355)) ([2345728](https://github.com/telekom/scale/commit/23457285b6d07ba94647c139bf5276cae801f94f))
+
+
+
+
+
+# [3.0.0-beta.9](https://github.com/telekom/scale/compare/v3.0.0-beta.8...v3.0.0-beta.9) (2021-05-17)
+
+**Note:** Version bump only for package @telekom/scale-visual-tests
+
+
+
+
+
+# [3.0.0-beta.8](https://github.com/telekom/scale/compare/v3.0.0-beta.7...v3.0.0-beta.8) (2021-05-12)
+
+**Note:** Version bump only for package @telekom/scale-visual-tests
+
+
+
+
+
+# [3.0.0-beta.7](https://github.com/telekom/scale/compare/v3.0.0-beta.6...v3.0.0-beta.7) (2021-05-06)
+
+
+### Features
+
+* **text-field:** styled read-only attribute w/ tokens ([#313](https://github.com/telekom/scale/issues/313)) ([1abbdd6](https://github.com/telekom/scale/commit/1abbdd6c2bbda810ab3529026d0f837e649ae792))
+
+
+
+
+
+# [3.0.0-beta.6](https://github.com/telekom/scale/compare/v3.0.0-beta.5...v3.0.0-beta.6) (2021-05-03)
+
+
+### Features
+
+* **footer:** dynamic copyright text ([#318](https://github.com/telekom/scale/issues/318)) ([04d1ad3](https://github.com/telekom/scale/commit/04d1ad3b247159e10a51db88ce56e658b42147ed))
+
+
+
+
+
 # [3.0.0-beta.5](https://github.com/telekom/scale/compare/v3.0.0-beta.4...v3.0.0-beta.5) (2021-04-28)
 
 **Note:** Version bump only for package @telekom/scale-visual-tests